add config test for spleen
